workers/docling-worker/requirements.txt
docling
boto3
psycopg2-binary